+++ CHANGELOG.md
@@ -17,11 +17,16 @@
 
  * **@dlr-eoc/map-maplibre:**
   - Adding support for mapState viewAngle and rotation 
+  - Added new `UKIS_METADATA.ignoreOpacity` and `UKIS_METADATA.ignoreVisibility` which can be used in `CustomLayer` to prevent changing the opacity or visibility for some layers.
+  - Added new function `updateStyleLayerProperties(map: glMap, mllayer: TypedStyleLayer, layer: ukisLayer)` to update `paint`, `layout`, `maxzoom`, `minzoom`, `filter`, `source` and `source-layer` from ukisLayers.
 
 ### Bug Fixes
  * **@dlr-eoc/map-maplibre:**
   - Fix wrong coordinate order from `getExtent` [#216](7ff01801fd3af594aaa34b2479e921be88894dbe)
-  - Rename `set/getBearing` to `set/getRotation` because we want to have the same behavior as in openlayers and `set/getBearing` is expected differently in mapliebre
+  - Rename `set/getBearing` to `set/getRotation` because we want to have the same behavior as in openlayers and `set/getBearing` is expected differently in maplibre.
+  - Fix create `StyleLayer` id in `createCustomLayer` and `createStackedLayer`.
+  - Fix create default `StyleLayers` for `GeoJson` Data (test for `MultiPolygon` and `MultiLineString`).
+  - Do not modify referenced objects when creating `StyleLayers` from `CustomLayer` or `VectorLayer.options.style` see also (cddf2fc6a83444851914ed6619bc8d5ada8ce485)
 
 * **@dlr-eoc/map-cesium:**
   - Fix get entity.name for infoBox in none GeoJson layers